backend/requirements.txt
Flask
Flask-SocketIO
Flask-Cors
Flask-Bcrypt
pandas
pymongo
numpy
python-dotenv
python-socketio
scikit-learn
